+++ core-ng-test/src/test/java/core/framework/test/scheduler/TestJob.java
@@ -1,13 +1,14 @@
 package core.framework.test.scheduler;
 
 import core.framework.scheduler.Job;
+import core.framework.scheduler.JobContext;
 
 /**
  * @author neo
  */
 public class TestJob implements Job {
     @Override
-    public void execute() {
+    public void execute(JobContext context) {
 
     }
 }