+++ core-ng/src/main/java/core/framework/scheduler/JobContext.java
@@ -1,0 +1,16 @@
+package core.framework.scheduler;
+
+import java.time.ZonedDateTime;
+
+/**
+ * @author neo
+ */
+public class JobContext {
+    public final String name;
+    public final ZonedDateTime scheduledTime;
+
+    public JobContext(String name, ZonedDateTime scheduledTime) {
+        this.name = name;
+        this.scheduledTime = scheduledTime;
+    }
+}